Fixes an issue when you pass an url directly.
This happens for example when you fetch og.likes with
a next URL which is already with all the necessary fields in it.
With this fix you're able to pass a fully constructed URL (so,
in this case, it's merely a wrapper around `request` but it
helps when you can use the same API).